+++ libs/server-sdk/src/evaluation/evaluation_error.hpp
@@ -2,7 +2,6 @@
 
 #include <optional>
 #include <ostream>
-#include <optional>
 
 namespace launchdarkly::server_side::evaluation {